+++ odoo/local-src/bso_api/controllers/employee.py
@@ -32,17 +32,23 @@ class EmployeeController(http.Controller):
             'family_id',
             'calendar_id',
             'notes',
+            'entry_date',
+            'is_contractor',
             # HR Settings
             'company_id',
             'user_login',
-            'crt_date_start',
-            'user_id'
+            'user_id',
+            # basic fields
+            'id',
+            'create_date',
+            'write_date',
+            'display_name',
         ]
         employees = request.env['hr.employee'].sudo().search_read(
             [], fields=fields, offset=int(offset),
             limit=int(limit) if limit else None)
         for e in employees:
-            user_id = e['user_id'][0] if e['user_id'] else False
+            user_id = e['user_id'][0] if e.get('user_id') else False
             user = request.env['res.users'].sudo().browse(user_id)
             e['has_odoo_login'] = user.active
         return _response(employees)